+++ include/ConfigFile.hpp
@@ -14,7 +14,7 @@ public:
 
   Chameleon const& Value(std::string const& section, std::string const& entry) const;
 
-  Chameleon const& Value(std::string const& section, std::string const& entry, double value);
+  Chameleon const& Value(std::string const& section, std::string const& entry, float value);
   Chameleon const& Value(std::string const& section, std::string const& entry, std::string const& value);
 };
